Add vote for task, only user who have created task can delete them

+++ src/Controller/TaskController.php
@@ -81,6 +81,8 @@ class TaskController extends AbstractController
     #[Route('/tasks/{id}/delete', name: 'task_delete', methods: ['GET'])]
     public function deleteTaskAction(Task $task)
     {
+        $this->denyAccessUnlessGranted('delete', $task);
+        
         $this->em->remove($task);
         $this->em->flush();
